NepomukTrack::score() now returns 0

+++ src/core-impl/collections/nepomukcollection/meta/NepomukTrack.cpp
@@ -207,7 +207,7 @@ NepomukTrack::score() const
     // ask dr_lepper
     // seriously no clue
 
-    return double( 10 );
+    return double( 0 );
 }
 
 void
